Clean up orientation handling.
This (attempts to) fixe a bug where the initial orientation of a launched application
may not be correct, as well as fixing a lot of oddness/redundancy in doing things.

Orientation is now just stored inside LauncherApp, LauncherWindow retrieves it
from there and listens to changes via orientationChanged() signal.

Note that initial orientation will still be wrong, thanks to a Mobility bug:
QTMOBILITY-1385

Signed-off-by: Robin Burchell <viroteck@viroteck.net>