Support for applauncherd boosters
Details: This change splits application initialization in two
parts. The common part does not depend on the name of the application,
and can be run before knowing which application will be started. Slow
things like LauncherApp and LauncherWindow instantiations are made at
this phase. The application-specific part parses the command line
arguments and does the rest of the initializations before launching
the application.