Don't wrap LauncherWindow in another QWidget, subclass QDeclarativeView directly.
This removes a layer of indirection (window is asked to draw, asks QDV to draw, asks content to draw -
pretty silly) so that the QDV is asked to redraw content directly.

In theory, this might mean slightly lower latency (for redraw/events) and should mean lower memory usage
as there won't be a backing store for a useless widget.

Signed-off-by: Robin Burchell <viroteck@viroteck.net>