Fix BMC #19358
Add support for application splash screen by piggy backing on top of
the existing booster preload mechanism.  With this approach the preloaded
QML will immediatly a splash image provided by the mainWindow, which will
either be ~/.cache/APPNAME/splash.png if the file exist and the app was
started with the 'restore' cmd, or the application icon.