Fix BMC #21220
'app' is to specify both path where to find launched binary or qml files
and application name which is a unique name to guarantee application
singleton. But this prevents one binary running as different applications.

To solve this, 'app' and new-added 'appname' parameters are used to
specify application binary path and application name respectively.
If 'appname' is not set, to fall back with previous behavior, still use
'app' as application name.